+++ src/formats/utility.c
@@ -2,19 +2,19 @@
 #include <stdlib.h>
 #include "../uByte/uByte.h"
 
-void addOffset(unsigned long long int ***offsets, unsigned long long int *offsetsSize, unsigned long long int start, unsigned long long int end) {
+void addOffset(uint64_t ***offsets, uint64_t *offsetsSize, uint64_t start, uint64_t end) {
 	(*offsetsSize)++;
 	if (!(*offsets)) {
-		*offsets = (unsigned long long int**)malloc((*offsetsSize) * sizeof(unsigned long long int*));
+		*offsets = (uint64_t**)malloc((*offsetsSize) * sizeof(uint64_t*));
 		if (!(*offsets))
 			return;
 	} else {
-		*offsets = (unsigned long long int**)realloc(*offsets, (*offsetsSize) * sizeof(unsigned long long int*));
+		*offsets = (uint64_t**)realloc(*offsets, (*offsetsSize) * sizeof(uint64_t*));
 		if (!(*offsets))
 			return;
 	}
-	
-	(*offsets)[(*offsetsSize) - 1] = (unsigned long long int*)calloc(2, sizeof(unsigned long long int));
+
+	(*offsets)[(*offsetsSize) - 1] = (uint64_t*)calloc(2, sizeof(uint64_t));
 	if (!(*offsets)[(*offsetsSize) - 1])
 		return;
 	(*offsets)[(*offsetsSize) - 1][0] = start;
@@ -22,25 +22,25 @@ void addOffset(unsigned long long int ***offsets, unsigned long long int *offset
 	return;
 }
 
-void setOffset(unsigned long long int ***offsets, unsigned long long int *offsetsSize, unsigned long long int index, unsigned long long int newOffset) {
+void setOffset(uint64_t ***offsets, uint64_t *offsetsSize, uint64_t index, uint64_t newOffset) {
 	if (*offsets && (*offsets)[(*offsetsSize) - 1] && (index == 0 || index == 1))
 		(*offsets)[(*offsetsSize) - 1][index] = newOffset;
 }
 
-void ltob(unsigned int *little) {
+void ltob(uint32_t *little) {
 	uByte endianness;
 	endianness.byte = 1;
 	if (endianness.bits.bit7 != 1) {
-		unsigned int big = *little;
+		uint32_t big = *little;
 		*little = ((big & 0x000000ff) << 24u) | ((big & 0x0000ff00) << 8u) | ((big & 0x00ff0000) >> 8u) | ((big & 0xff000000) >> 24u);
 	}
 }
 
-void btol(unsigned int *big) {
+void btol(uint32_t *big) {
 	uByte endianness;
 	endianness.byte = 1;
 	if (endianness.bits.bit1 != 1) {
-		unsigned int little = *big;
+		uint32_t little = *big;
 		*big = ((little & 0x000000ff) << 24u) | ((little & 0x0000ff00) << 8u) | ((little & 0x00ff0000) >> 8u) | ((little & 0xff000000) >> 24u);
 	}
 }
